requirements.txt
langgraph
langchain
ipykernel
python-dotenv
langchain-tavily



uvicorn
fastapi
python-multipart